+++ mopidy/__main__.py
@@ -40,11 +40,13 @@ def main():
     signal.signal(signal.SIGUSR1, pykka.debug.log_thread_tracebacks)
 
     try:
+        registry = ext.Registry()
+
         root_cmd = commands.RootCommand()
         config_cmd = commands.ConfigCommand()
         deps_cmd = commands.DepsCommand()
 
-        root_cmd.set(extension=None)
+        root_cmd.set(extension=None, registry=registry)
         root_cmd.add_child('config', config_cmd)
         root_cmd.add_child('deps', deps_cmd)
 
@@ -84,7 +86,6 @@ def main():
                 enabled_extensions.append(extension)
 
         log_extension_info(installed_extensions, enabled_extensions)
-        ext.register_gstreamer_elements(enabled_extensions)
 
         # Config and deps commands are simply special cased for now.
         if args.command == config_cmd:
@@ -108,10 +109,13 @@ def main():
                 args.extension.ext_name)
             return 1
 
+        for extension in enabled_extensions:
+            extension.setup(registry)
+
         # Anything that wants to exit after this point must use
         # mopidy.utils.process.exit_process as actors can have been started.
         try:
-            return args.command.run(args, proxied_config, enabled_extensions)
+            return args.command.run(args, proxied_config)
         except NotImplementedError:
             print(root_cmd.format_help())
             return 1