+++ tests/backends/local/tracklist_test.py
@@ -18,7 +18,9 @@ class LocalTracklistProviderTest(unittest.TestCase):
     config = {
         'local': {
             'media_dir': path_to_data_dir(''),
+            'data_dir': path_to_data_dir(''),
             'playlists_dir': b'',
+            'library': 'json',
         }
     }
     tracks = [